requirements.txt
flask==3.0.3
openai==1.72.0
spacy==3.8.7
smart-open==5.2.1
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
markupsafe==3.0.2
nltk==3.9.1
opencv-python==4.11.0.86
pandas==2.2.3
pdfplumber==0.11.6
pillow==11.1.0
psycopg2-binary==2.9.10
pytesseract==0.3.13
scikit-learn==1.6.1
sqlalchemy==2.0.40
torch==2.6.0
trafilatura==2.0.0
werkzeug==3.0.3
email-validator==2.2.0
flask-login==0.6.3
pyotp==2.9.0
flask-wtf==1.2.1
